+++ changelog.md
@@ -6,6 +6,7 @@
 + Remove "Coordinate" name alignment to "Coord".
 + Follows PEP517.
 + New submodule "optimization".
++ Rename module "planar_linkage" as "planar".
 
 # Version 20.12